update error test

+++ tests/errors/semantic/blocking/EXTEND_LIST_INCOMPATIBLE_TYPES_0.py
@@ -2,6 +2,6 @@
 
 import numpy as np
 a = [1,2,3]
-b = np.array([4, 5, 6], dtype=np.int32)
+b = np.array([4, 5, 6], dtype=np.int16)
 a.extend(b)
